app/backend/functions/requirements.txt
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Deprecated==1.2.3
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.6
futures==3.1.1
google-api-core==1.5.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
googleapis-common-protos==1.5.3
grpcio==1.16.0
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyGithub==1.43.2
PyJWT==1.6.4
pytz==2018.6
requests==2.20.0
rsa==4.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
wrapt==1.10.11
